split up contact form code a little
