+++ phpstan.neon.dist
@@ -52,6 +52,8 @@ parameters:
         - %currentWorkingDirectory%/modules/contact-form/module-cpt.php
         - %currentWorkingDirectory%/modules/contact-form/module-cron.php
         - %currentWorkingDirectory%/modules/contact-form/module-ajax.php
+        - %currentWorkingDirectory%/modules/contact-form/module-admin.php
+        - %currentWorkingDirectory%/modules/contact-form/module-submit-form.php
     ignoreErrors:
         # Uses func_get_args()
         - '#^Function [a-zA-Z_]+ invoked with [1-9] parameters?, [0-9\-]+ required\.$#'
